fix: Fix the create bug

+++ app/crud.py
@@ -2,7 +2,7 @@
 
 from sqlalchemy.orm import Session
 from app.models import Item
-from app.schemas import ItemCreate
+from app.schemas import ItemCreate, ItemUpdate
 
 def get_item(db:Session, item_id: int):
     return db.query(Item).filter(Item.id == item_id).first()
@@ -10,19 +10,22 @@ def get_item(db:Session, item_id: int):
 def create_item(db:Session, item:ItemCreate):
     db_item = Item(name = item.name, description=item.description)
     db.add(db_item)
-    db.commit
+    db.commit()
     db.refresh(db_item)
     return db_item
 
-def update_item(db:Session, item_id:int, item:ItemCreate):
+def update_item(db:Session, item_id:int, item:ItemUpdate):
     db_item = db.query(Item).filter(Item.db == item_id).first()
-    db_item.name = item.name
-    db_item.description = item.description
-    db.commit()
+    if db_item:
+        db_item.name = item.name
+        db_item.description = item.description
+        db.commit()
+        db.refresh(db_item)
     return db_item
 
 def delete_item(db:Session, item_id: int):
     db_item = db.query(Item).filter(Item.id == item_id).first()
-    db.delete(db_item)
-    db.commit()
+    if db_item:
+        db.delete(db_item)
+        db.commit()
     return db_item
